requirements.txt
pymilvus
"pymilvus[model]"
python-nmap
openai
ollama
